+++ tensorflow_addons/custom_ops/README.md
@@ -6,4 +6,4 @@
 | Image | Ops for image manipulation   |
 | Seq2seq | Ops for seq2seq encoder-decoder framework |
 | Text |  Ops for text processing  |
-| OpticalFlow |  Ops for optical flow processing  |
+| Layers |  Ops for model layers  |
